+++ .vscode/settings.json
@@ -2,6 +2,7 @@
   "files.associations": {
     "sstream": "cpp",
     "iostream": "cpp",
-    "queue": "cpp"
+    "queue": "cpp",
+    "set": "cpp"
   }
 }
